+++ Libplanet.RocksDBStore.Tests/MonoRocksDBStoreTest.cs
@@ -3,10 +3,10 @@ using System.IO;
 using Libplanet.Blockchain;
 using Libplanet.Blockchain.Policies;
 using Libplanet.Store;
+using Libplanet.Store.Trie;
 using Libplanet.Tests.Blockchain;
 using Libplanet.Tests.Common.Action;
 using Libplanet.Tests.Store;
-using Libplanet.Tests.Store.Trie;
 using Xunit;
 using Xunit.Abstractions;