+++ Libplanet.Tests/Action/ActionContextTest.cs
@@ -5,7 +5,6 @@ using Bencodex.Types;
 using Libplanet.Action;
 using Libplanet.Assets;
 using Libplanet.Store.Trie;
-using Libplanet.Tests.Store.Trie;
 using Libplanet.Tx;
 using Xunit;
 
@@ -149,7 +148,7 @@ namespace Libplanet.Tests.Action
         [InlineData(false)]
         public void LazyPreviousStateRootHash(bool callPreviousStateRootHash)
         {
-            var keyValueStore = new MemoryKeyValueStore();
+            IKeyValueStore keyValueStore = new MemoryKeyValueStore();
             ITrie previousBlockStatesTrie = new MerkleTrie(keyValueStore);
             previousBlockStatesTrie = previousBlockStatesTrie.Set(new byte[0], default(Null));
             var actionContext = new ActionContext(