+++ Libplanet.Tests/Net/Messages/NetMQMessageCodecTest.cs
@@ -9,8 +9,8 @@ using Libplanet.Crypto;
 using Libplanet.Net;
 using Libplanet.Net.Messages;
 using Libplanet.Store;
+using Libplanet.Store.Trie;
 using Libplanet.Tests.Common.Action;
-using Libplanet.Tests.Store.Trie;
 using NetMQ;
 using Xunit;
 
@@ -63,7 +63,7 @@ namespace Libplanet.Tests.Net.Messages
             IBlockPolicy<DumbAction> policy = new BlockPolicy<DumbAction>();
             BlockChain<DumbAction> chain = TestUtils.MakeBlockChain(
                 policy,
-                new DefaultStore(null),
+                new MemoryStore(),
                 new TrieStateStore(new MemoryKeyValueStore())
             );
             var codec = new Codec();