+++ Libplanet.Tests/Store/StoreTest.cs
@@ -932,6 +932,10 @@ namespace Libplanet.Tests.Store
             Assert.Equal(Fx.Block2.Hash, store.IndexBlockHash(chainB, 2));
             Assert.Equal(Fx.Block2.Hash, store.IndexBlockHash(chainC, 2));
             Assert.Equal(Fx.Block3.Hash, store.IndexBlockHash(chainC, 3));
+
+            Assert.Throws<ChainIdNotFoundException>(
+                () => store.ForkBlockIndexes(Guid.NewGuid(), Guid.NewGuid(), Fx.Block2.Hash)
+            );
         }
 
         [SkippableFact]
@@ -1074,6 +1078,14 @@ namespace Libplanet.Tests.Store
             Assert.Equal(1, store.GetTxNonce(destinationChainId, Fx.Address1));
             Assert.Equal(2, store.GetTxNonce(destinationChainId, Fx.Address2));
             Assert.Equal(3, store.GetTxNonce(destinationChainId, Fx.Address3));
+
+            store.IncreaseTxNonce(sourceChainId, Fx.Address1, 1);
+            Assert.Equal(2, store.GetTxNonce(sourceChainId, Fx.Address1));
+            Assert.Equal(1, store.GetTxNonce(destinationChainId, Fx.Address1));
+
+            Assert.Throws<ChainIdNotFoundException>(
+                () => store.ForkTxNonces(Guid.NewGuid(), Guid.NewGuid())
+            );
         }
 
         [SkippableFact]