+++ Libplanet/Store/DefaultStore.cs
@@ -239,6 +239,14 @@ namespace Libplanet.Store
             BlockHash branchpoint)
         {
             LiteCollection<HashDoc> srcColl = IndexCollection(sourceChainId);
+            if (!srcColl.Exists(_ => true))
+            {
+                throw new ChainIdNotFoundException(
+                    sourceChainId,
+                    $"No such chain ID: {sourceChainId}."
+                );
+            }
+
             LiteCollection<HashDoc> destColl = IndexCollection(destinationChainId);
 
             BlockHash? genesisHash = IterateIndexes(sourceChainId, 0, 1)
@@ -631,8 +639,15 @@ namespace Libplanet.Store
         public override void ForkTxNonces(Guid sourceChainId, Guid destinationChainId)
         {
             LiteCollection<BsonDocument> srcColl = TxNonceCollection(sourceChainId);
-            LiteCollection<BsonDocument> destColl = TxNonceCollection(destinationChainId);
+            if (!srcColl.Exists(_ => true))
+            {
+                throw new ChainIdNotFoundException(
+                    sourceChainId,
+                    $"No such chain ID: {sourceChainId}."
+                );
+            }
 
+            LiteCollection<BsonDocument> destColl = TxNonceCollection(destinationChainId);
             destColl.InsertBulk(srcColl.FindAll());
         }